implementation of Upload button.
Updated the menu option UI/UX. When user click the Register tab in menu it will direct into a new page which the user able to upload or go back to the menu. If user clicks the upload button it will direct into a new page that gives the user to access the files in their phone and upload a song. Register function is not yet implemented. Have to re-code the function if the user clicks the log out button, it should connect with the SignUp project.